created camera, functions to support it. makes it so the camera centers on the player in player.c, changed draw functions in entity and world to adjust to the camera offset, changed entity_update_position to not check the entity against world bounds directly inside that function, instead created entity_check_world_bounds, added camera setup in game.c

+++ include/physics.h
@@ -5,7 +5,7 @@
 #include "gfc_config.h"
 #include "gfc_shape.h"
 
-#define FGRAV ((GFC_Vector2D) {0, 0.2})
+#define FGRAV ((GFC_Vector2D) {(float) 0, (float) 0.2})
 
 typedef struct Physics_Object_s
 {
